+++ docs/changelog.md
@@ -9,6 +9,11 @@ All notable changes to DartLab will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.6] - 2026-04-03
+
+### Added
+- OAuth 수동 토큰 입력 (방화벽 환경용)
+
 ## [0.8.5] - 2026-04-03
 
 ### Added